pyproject.toml
[tool.poetry]
name = "configerelle"
version = "0.1.0"
description = ""
authors = ["NewWars <15692866+OriDevTeam@users.noreply.github.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
attrs = "^22.2.0"
click = "^8.1.3"
antlr4-python3-runtime = "^4.12.0"
flatdict = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
